move meta out of Layout/index
